requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
tkinter-tooltip==2.0.0
customtkinter>=5.2.0 